separado o componente
